+++ server/src/routes/api/v1/usersRouter.js
@@ -33,7 +33,6 @@ usersRouter.patch("/:id", uploadImage.single("image"), async (req, res) => {
     });
     return res.status(201).json({ user: updatedUser });
   } catch (error) {
-    console.log(error);
     return res.status(500).json({ errors: error });
   }
 });
